增加游戏记录接口 增加结算详情接口

+++ src/main/java/com/dx/facade/game/resp/GameOrderUserSettlementResp.java
@@ -21,5 +21,5 @@ public class GameOrderUserSettlementResp {
     @ApiModelProperty("总带入")
     private BigDecimal bringMidwayChip;
     @ApiModelProperty("总带入次数")
-    private BigDecimal bringCount;
+    private Long bringCount;
 }
